+++ src/utils/common.js
@@ -1,27 +1,5 @@
-import {MILISECONDS} from './const.js';
+import {MILISECONDS} from '../const.js';
 
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
 
 const castTimeFormat = (value) => {
   return value < 10 ? `0${value}` : String(value);
